requirements.txt
flask
gunicorn
python-dotenv
PyPDF2
langchain
langchain-core
langchain-classic
langchain-community
langchain-text-splitters
langchain-groq
faiss-cpu
sentence-transformers
huggingface-hub